requirements.txt
pandas
numpy
scikit-learn
matplotlib
gradio
joblib
groq
python-dotenv
plotly